pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plaso-downloader"
version = "0.1.0"
description = "CLI tool for downloading plaso classroom videos and PDFs."
readme = "README.md"
authors = [{ name = "Codex", email = "codex@example.com" }]
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "tqdm>=4.66.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.1",
    "pillow>=10.0.0",
]

[project.scripts]
plaso-downloader = "plaso_downloader.main:main"

[tool.setuptools]
packages = ["plaso_downloader"]
